pyproject.toml
[project]
name = "mcp-gsheets"
version = "0.1.0"
description = "Comprehensive MCP server for Google Sheets API - covering the v4 API including charts, formatting, pivot tables, and more"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.5.0",
    "google-auth>=2.28.1",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.117.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.9",
    "cryptography>=41.0.0",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
]

[[project.authors]]
name = "Erik Möller"
email = "emoller@opera.com"

[project.urls]
"Homepage" = "https://github.com/opera-emoller/mcp-gsheets"
"Bug Tracker" = "https://github.com/opera-emoller/mcp-gsheets/issues"
"Repository" = "https://github.com/opera-emoller/mcp-gsheets.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-gsheets = "mcp_gsheets:main"
mcp-gsheets-web = "mcp_gsheets.webapp:main"
mcp-gsheets-http = "mcp_gsheets.http_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_gsheets"]
